+++ CHANGELOG.md
@@ -1,5 +1,45 @@
 # Changelog
 
+### [0.5.2](https://www.github.com/licenseware/licenseware-sdk/compare/v0.5.1...v0.5.2) (2021-06-22)
+
+
+### Bug Fixes
+
+* AppCreator issues fixed ([9e37fef](https://www.github.com/licenseware/licenseware-sdk/commit/9e37fefdbf585fe3a251ed096ebb39037484cc2f))
+
+### [0.5.1](https://www.github.com/licenseware/licenseware-sdk/compare/v0.5.0...v0.5.1) (2021-06-22)
+
+
+### Bug Fixes
+
+* removed dependency clutter ([f270ab5](https://www.github.com/licenseware/licenseware-sdk/commit/f270ab5b142275bc3977956b39825b9c2e2ff2b0))
+
+## [0.5.0](https://www.github.com/licenseware/licenseware-sdk/compare/v0.4.1...v0.5.0) (2021-06-22)
+
+
+### Features
+
+* added AppCreator which allows register of uploder, history, quota, app_definition all in one go ([0aae36a](https://www.github.com/licenseware/licenseware-sdk/commit/0aae36ad015943f63d2920ee8d027f6ebad54d31))
+
+
+### Bug Fixes
+
+* revert to mongodata module to solve imports ([b536b9a](https://www.github.com/licenseware/licenseware-sdk/commit/b536b9a7c3e7a6c8bb2c5808022a605457df9a78))
+
+
+### Documentation
+
+* added docs for AppCreator ([62ddefc](https://www.github.com/licenseware/licenseware-sdk/commit/62ddefc17d44bdf553758ec926552b8289283bd6))
+
+### [0.4.1](https://www.github.com/licenseware/licenseware-sdk/compare/v0.4.0...v0.4.1) (2021-06-21)
+
+
+### Bug Fixes
+
+* auto refresh token before every request ([ba830d9](https://www.github.com/licenseware/licenseware-sdk/commit/ba830d9636083e78bf3965ca8b04303ab77aafc6))
+* fix relative import ([b9a561f](https://www.github.com/licenseware/licenseware-sdk/commit/b9a561f97d285c38f789e4550f1e10f3ef46e5b4))
+* fixed relative import for mongodata ([ceaa765](https://www.github.com/licenseware/licenseware-sdk/commit/ceaa7659fe5b83549df9bbebdb80d82bd4965e93))
+
 ## [0.4.0](https://www.github.com/licenseware/licenseware-sdk/compare/v0.3.4...v0.4.0) (2021-06-18)